requirements.txt
pytest==8.3.5
selenium==4.29.0
allure-pytest==2.13.5
python-dotenv==1.1.0
